Improve support for different rst code block syntaxes
This adds support for all reStructuredText code block syntaxes:

- .. code-block::
- .. code-block:: language
- Example::
- ::

The previously used regular expression approach is simplified. Regular
expression searches are only used to find the basic syntax elements.
A new search step finds ‘double colon’ code block syntax in addition
to the ‘code-block::’ syntax.

The logic to figure out the language (now with a default fallback),
and to detect the boundaries of the block to highlight, is now
explicitly written out in elisp, using a helper from the built-in
rst.el. This improves detection for indented blocks. An internal
helper is extracted from the previous code, so that it can be used
for for all code-block syntaxes.

Added a dependency on s.el to use its string functions, and added
more (require) statements to load all used functions.

Added a sample file to aid developing and debugging.

While at it, fix all linter warnings.

Closes #4.